refactorización del código (creación de nuevas views y cambios en ambos archivos urls)
